requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
tensorflow==2.16.1
scikit-learn==1.5.0
pandas==2.1.2
numpy==1.26.1
nltk==3.8.1
matplotlib==3.10.1